Do not create DB connection in a context manager
@mcdonnnj pointed out that psycopg.connect() does not entirely act as
expected in that it does not automatically close the connection when
you exit the context.  It only rolls back transactions in the event
that one fails.

Since we are managing transactions in this code via explicit calls to
psycopg.Connection.commit(), the context manager isn't actually buying
us anything; therefore, I have gotten rid of it.

I added an explicit psycopg.Connection.close() at the bottom of the
loop.  I don't want to create the connection outside the loop and
leave it open because the guacscanner may be up and running for months
at a time, and who knows what could go wrong.  It seems safer to just
recreate the connection at the top of the loop.