Use a connection attribute to hold the instance ID
Use this attribute as a unique identifier for the connection instead
of the connection name.  The connection name now contains the various
IP addresses associated with the instance, so it could change without
the actual instance ID changing.

This is a better way to identify the connections associated with a
particular instance ID anyway.